исправил мелкие ошибки

+++ ind_2.2.py
@@ -10,7 +10,7 @@
 import sys
 
 if __name__ == '__main__':
-    A = tuple(map(int, input().split()))
+    A = tuple(map(float, input().split()))
     if not A:
         print('Заданный кортеж пуст', file=sys.stderr)
         exit(1)
